Switch to using a simpler-but-still-hacky way of invoking stata
We create a wrapper script around the supplied script that writes a file
if it successfully runs the script. We use the contents of this file as
indication of success or failure, and exit approriately.

We can then use the regular stata (without -b), which gives stdout much
more simply.

Also, add an experimental features to auto import ado files in the
./libraries directory of a study.