Script to generate the pyarrow test fixtures
This generates two arrow files, one with 1000 observations, which is
read in one batch, and another with 100,000 observations, which is
read in two batches.

It creates a variable, with and without nulls (except for bools), for
each type we expect to encounter in an arrow file. The values are
random, except for rows 0, 1 and -1, which have fixed values inserted
so we can check them in the tests.

For the multi-batch data, the interger-type values are manipulated so
that the first batch always gets small (byte-sized) numbers only, which
will need to be adjusted on the second batch to ensure the variables and
replaced null values are of the right types.