Exit with the standard stata "dataset in memory has changed"
if any variables already exist in the stata dataset. This
can happen quite often, if a user tries to load another file
without clearing the dataset, or if variable have been generated
some other way. This follows the same behaviour as other commands
such as 'import delimited' so shouldn't be surprising to stata
users.